my_web/requirements.txt
Click==7.0
Django==2.2.1
django-ckeditor==5.7.1
django-js-asset==1.2.2
django-model-utils==3.1.2
django-mptt==0.10.0
django-notifications-hq==1.5.0
django-password-reset==2.0
django-taggit==1.1.0
Flask==1.0.2
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10
jsonfield==2.0.2
Markdown==3.1.1
markdown2==2.3.7
MarkupSafe==1.1.1
Pillow==6.0.0
psycopg2-binary==2.8.2
Pygments==2.4.0
pytz==2019.1
sqlparse==0.3.0
Werkzeug==0.15.1
WTForms==2.2.1
